refactor: consolidate _env_loader/_install_guide/_optional_deps into scitex.helpers/
Group internal helper modules (env loader, install guidance, optional-deps
checks) under scitex.helpers/ for a clearer top-level API. Backwards-compat
shims at the original module paths re-export every public + private symbol so
legacy `from scitex._env_loader import X` calls keep working.
